Xhr: Активация http-кэша вместе с rewriteUrl в выдаче.
От IndexedDB пока полный отказ, т.к. требует отдельного разрешения на
экране. Реализация выполнена через однократное кэширование без
карты кэша.

MGeoPoint: широта и долгота теперь BigDecimal.
Точное нормирование десятичных частей широты и долготы.
Укорачивание и нормализация ссылок.
js-кэширование в округлённых координатах.

Не ясно, нормально ли работает кэш, т.к. он только https, не для
локалхоста.

+++ src1/server/geo/mgeo/src/main/scala/io/suggest/geo/CircleGsJvm.scala
@@ -43,7 +43,7 @@ object CircleGsJvm extends GsStaticJvmQuerable {
 
   override def toEsShapeBuilder(gs: Shape_t) = {
     ShapeBuilders.newCircleBuilder()
-      .center(gs.center.lon, gs.center.lat)
+      .center(gs.center.lon.doubleValue(), gs.center.lat.doubleValue())
       .radius(gs.radiusM, DistanceUnit.METERS)
   }
